upgrade circe version

+++ src/main/scala/com/zio/examples/http4s_doobie/Data.scala
@@ -2,4 +2,4 @@ package com.zio.examples.http4s_doobie
 
 final case class User(id: Long, name: String)
 
-final case class UserNotFound(id: Int) extends Exception
+final case class UserNotFound(id: Long) extends Exception
